+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
 ## Version history
-* 0.0.4 - Associate voorbeeld_type tax with Instrument CPT. Remove archive options.
+* 1.0.1 - Better filter breadcrumb list when viewing a single Voorbeeld CPT.
+* 1.0.0 - Launch Toolkit Taal site!
 * 0.0.3 - Changed CPT icon to 'dashicons-media-document'
 * 0.0.2 - Custom post type 'voorbeeld' added.
 * 0.0.1 - Initial version.